requirements.txt
pandas
requests
streamlit
folium
streamlit-folium
rasterio
shapely
geopy
matplotlib
numpy
pyyaml
pystac-client
planetary-computer
stackstac
rioxarray
xarray
dask
distributed
reportlab
# --- AI & Agent Dependencies ---
langchain
langchain-groq
langchain-community
langchain-huggingface
langchain-chroma
langgraph
chromadb
sentence-transformers
python-dotenv
pypdf